# apps/command/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Every error response: ``{\"error\": {\"code\", \"message\", \"run_id\"}}``.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RainNowcast": {
        "additionalProperties": false,
        "description": "The cycle's rain averaged over the city: the time bar's spread band (SPEC.md 7.2).\n\n``members`` is the AOI-mean hyetograph of each member - the ensemble's own disagreement,\nnot a smoothed envelope - and ``steps`` is the quantile band across those members, taken\nwith the same quantiles and interpolation as the per-pixel products so the band and the map\ncannot mean different things. Note what the band is *of*: the spread of the AOI mean, which\nis narrower than the spread over any one street.",
        "properties": {
          "bundle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Replay bundle id, null when live.",
            "title": "Bundle"
          },
          "city": {
            "description": "City slug the AOI belongs to, e.g. mumbai.",
            "title": "City",
            "type": "string"
          },
          "members": {
            "description": "One hyetograph per member, in member order.",
            "items": {
              "$ref": "#/components/schemas/RainMemberSeries"
            },
            "title": "Members",
            "type": "array"
          },
          "mode": {
            "description": "baked = read from a run under data/runs; live = computed for this request.",
            "enum": [
              "baked",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "n_members": {
            "description": "Ensemble members (20 in the prototype).",
            "minimum": 1.0,
            "title": "N Members",
            "type": "integer"
          },
          "n_steps": {
            "description": "Forecast steps (36 = 3 h at 5 min).",
            "minimum": 1.0,
            "title": "N Steps",
            "type": "integer"
          },
          "notes": {
            "description": "The run's honesty labels, verbatim: which Z-R relation ran, what the gauge merge did, whether the NWP blend is off (SPEC.md 0.6). The console prints them as they arrive.",
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "nowcaster": {
            "anyOf": [
              {
                "enum": [
                  "pysteps_steps",
                  "fallback_steps"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nowcaster that produced the ensemble. Null when the run's rain/cube.zarr is not on disk to say, which is the only artifact that records it.",
            "title": "Nowcaster"
          },
          "peak_p50_mm_h": {
            "description": "Highest median rain rate over the horizon; 0 when the series is empty.",
            "readOnly": true,
            "title": "Peak P50 Mm H",
            "type": "number"
          },
          "peak_ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When :attr:`peak_p50_mm_h` occurs, or null when the series is empty.",
            "readOnly": true,
            "title": "Peak Ts"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run the products were read from, or null when they were computed on demand and not published as a run.",
            "title": "Run Id"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seed the ensemble was drawn with.",
            "title": "Seed"
          },
          "stage_ms": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Wall-clock per stage in ms, for the cycle budget bar (SPEC.md 7.2).",
            "title": "Stage Ms",
            "type": "object"
          },
          "step_min": {
            "description": "Minutes between forecast steps.",
            "exclusiveMinimum": 0.0,
            "title": "Step Min",
            "type": "number"
          },
          "steps": {
            "description": "One entry per forecast step, in time order.",
            "items": {
              "$ref": "#/components/schemas/RainStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "valid_ts": {
            "description": "Cycle time of the run (IST): the instant the forecast was made. Each step's own valid time is on the step.",
            "format": "date-time",
            "title": "Valid Ts",
            "type": "string"
          },
          "zr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ZRRelation"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Z-R relation of this cycle."
          }
        },
        "required": [
          "run_id",
          "valid_ts",
          "mode",
          "city",
          "n_members",
          "n_steps",
          "step_min",
          "peak_p50_mm_h",
          "peak_ts"
        ],
        "title": "RainNowcast",
        "type": "object"
      },
...
      "RainPointSeries": {
        "additionalProperties": false,
        "description": "The rain fan chart at one point (SPEC.md 7.2, task P3.8).\n\nSampled from the 500 m Sky grid, which is the grid the quantiles are published on: the\n30 m AOI resample is the Twin's forcing and is never written to disk (SPEC.md 10.3), so a\npoint series describes the pixel the junction sits in, not the junction's own square metre.",
        "properties": {
          "bundle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Replay bundle id, null when live.",
            "title": "Bundle"
          },
          "city": {
            "description": "City slug the AOI belongs to, e.g. mumbai.",
            "title": "City",
            "type": "string"
          },
          "exceedance_mm_h": {
            "description": "The two thresholds p_gt_20 and p_gt_40 report, in mm/h.",
            "items": {
              "type": "number"
            },
            "maxItems": 2,
            "minItems": 2,
            "title": "Exceedance Mm H",
            "type": "array"
          },
          "mode": {
            "description": "baked = read from a run under data/runs; live = computed for this request.",
            "enum": [
              "baked",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "n_members": {
            "description": "Ensemble members (20 in the prototype).",
            "minimum": 1.0,
            "title": "N Members",
            "type": "integer"
          },
          "n_steps": {
            "description": "Forecast steps (36 = 3 h at 5 min).",
            "minimum": 1.0,
            "title": "N Steps",
            "type": "integer"
          },
          "notes": {
            "description": "The run's honesty labels, verbatim: which Z-R relation ran, what the gauge merge did, whether the NWP blend is off (SPEC.md 0.6). The console prints them as they arrive.",
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "nowcaster": {
            "anyOf": [
              {
                "enum": [
                  "pysteps_steps",
                  "fallback_steps"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nowcaster that produced the ensemble. Null when the run's rain/cube.zarr is not on disk to say, which is the only artifact that records it.",
            "title": "Nowcaster"
          },
          "peak_p50_mm_h": {
            "description": "Highest median rain rate over the horizon; 0 when the series is empty.",
            "readOnly": true,
            "title": "Peak P50 Mm H",
            "type": "number"
          },
          "peak_ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When :attr:`peak_p50_mm_h` occurs, or null when the series is empty.",
            "readOnly": true,
            "title": "Peak Ts"
          },
          "point": {
            "$ref": "#/components/schemas/RainPoint"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run the products were read from, or null when they were computed on demand and not published as a run.",
            "title": "Run Id"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seed the ensemble was drawn with.",
            "title": "Seed"
          },
          "stage_ms": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Wall-clock per stage in ms, for the cycle budget bar (SPEC.md 7.2).",
            "title": "Stage Ms",
            "type": "object"
          },
          "step_min": {
            "description": "Minutes between forecast steps.",
            "exclusiveMinimum": 0.0,
            "title": "Step Min",
            "type": "number"
          },
          "steps": {
            "description": "One entry per forecast step, in time order.",
            "items": {
              "$ref": "#/components/schemas/RainPointStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "valid_ts": {
            "description": "Cycle time of the run (IST): the instant the forecast was made. Each step's own valid time is on the step.",
            "format": "date-time",
            "title": "Valid Ts",
            "type": "string"
          },
          "zr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ZRRelation"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Z-R relation of this cycle."
          }
        },
        "required": [
          "run_id",
          "valid_ts",
          "mode",
          "city",
          "n_members",
          "n_steps",
          "step_min",
          "point",
          "exceedance_mm_h",
          "peak_p50_mm_h",
          "peak_ts"
        ],
        "title": "RainPointSeries",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Team VIT, SIH 2026 PS SIH26085"
    },
    "description": "Street-level urban flood nowcasting digital twin (SIH 2026, PS SIH26085, MoES). Every response carries `run_id` and `valid_ts` where a run is involved; errors use `{ error: { code, message, run_id } }`. Endpoints answering 501 are contract-only until their engine's phase lands.",
    "title": "VARUNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/city/{city}/basemap.pmtiles": {
      "get": {
        "description": "The city's PMTiles basemap, whole or by byte range.",
        "operationId": "city_basemap_v1_city__city__basemap_pmtiles_get",
        "parameters": [
          {
            "in": "path",
            "name": "city",
            "required": true,
            "schema": {
              "title": "City",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.pmtiles": {}
            },
            "description": "PMTiles v3 archive"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Basemap not built for this city"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offline vector basemap (OSM roads, buildings, waterways; WorldCover water)",
        "tags": [
          "basemap"
        ]
      }
    },
...
    "/v1/nowcast/rain": {
      "get": {
        "description": "The time bar's spread band (SPEC.md 7.2): the ensemble's own disagreement about how\nmuch rain the city is about to get, one line per member over the three-hour horizon.",
        "operationId": "nowcast_rain_v1_nowcast_rain_get",
        "parameters": [
          {
            "description": "Run id; default = the newest run under data/runs that carries rain.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Run id; default = the newest run under data/runs that carries rain.",
              "title": "Run Id"
            }
          },
          {
            "description": "Compute this cycle from the replay bundle instead of reading a baked run. The result is not published as a run, and the response reads mode=live.",
            "in": "query",
            "name": "compute",
            "required": false,
            "schema": {
              "default": false,
              "description": "Compute this cycle from the replay bundle instead of reading a baked run. The result is not published as a run, and the response reads mode=live.",
              "title": "Compute",
              "type": "boolean"
            }
          },
          {
            "description": "Bundle to compute from; default = the open replay, else VARUNA_BUNDLE.",
            "in": "query",
            "name": "bundle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bundle to compute from; default = the open replay, else VARUNA_BUNDLE.",
              "title": "Bundle"
            }
          },
          {
            "description": "Cycle time to compute (ISO 8601 with offset, e.g. 2019-07-02T06:40:00+05:30). Floored onto the bundle's five-minute cycle ladder; default = the replay clock.",
            "in": "query",
            "name": "t",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle time to compute (ISO 8601 with offset, e.g. 2019-07-02T06:40:00+05:30). Floored onto the bundle's five-minute cycle ladder; default = the replay clock.",
              "title": "T"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RainNowcast"
                }
              }
            },
            "description": "AOI-mean rain per member in mm/h, with the p10/p50/p90 band"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No run, bundle or city to answer from"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request names a point or cycle that cannot be served"
          }
        },
        "summary": "Rain over the city: every member's hyetograph and the spread band",
        "tags": [
          "nowcast"
        ]
      }
    },
    "/v1/nowcast/rain/series": {
      "get": {
        "description": "The fan chart at Hindmata (task P3.8), or at any point inside the radar domain.",
        "operationId": "nowcast_rain_series_v1_nowcast_rain_series_get",
        "parameters": [
          {
            "description": "Register id, slug or part of a name from the city's hotspots, e.g. Hindmata. The coordinate comes from the register, with the source it was verified against.",
            "in": "query",
            "name": "hotspot",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Register id, slug or part of a name from the city's hotspots, e.g. Hindmata. The coordinate comes from the register, with the source it was verified against.",
              "title": "Hotspot"
            }
          },
          {
            "description": "WGS84 longitude.",
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "WGS84 longitude.",
              "title": "Lon"
            }
          },
          {
            "description": "WGS84 latitude.",
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "WGS84 latitude.",
              "title": "Lat"
            }
          },
          {
            "description": "Run id; default = the newest run under data/runs that carries rain.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Run id; default = the newest run under data/runs that carries rain.",
              "title": "Run Id"
            }
          },
          {
            "description": "Compute this cycle from the replay bundle instead of reading a baked run. The result is not published as a run, and the response reads mode=live.",
            "in": "query",
            "name": "compute",
            "required": false,
            "schema": {
              "default": false,
              "description": "Compute this cycle from the replay bundle instead of reading a baked run. The result is not published as a run, and the response reads mode=live.",
              "title": "Compute",
              "type": "boolean"
            }
          },
          {
            "description": "Bundle to compute from; default = the open replay, else VARUNA_BUNDLE.",
            "in": "query",
            "name": "bundle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bundle to compute from; default = the open replay, else VARUNA_BUNDLE.",
              "title": "Bundle"
            }
          },
          {
            "description": "Cycle time to compute (ISO 8601 with offset, e.g. 2019-07-02T06:40:00+05:30). Floored onto the bundle's five-minute cycle ladder; default = the replay clock.",
            "in": "query",
            "name": "t",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cycle time to compute (ISO 8601 with offset, e.g. 2019-07-02T06:40:00+05:30). Floored onto the bundle's five-minute cycle ladder; default = the replay clock.",
              "title": "T"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RainPointSeries"
                }
              }
            },
            "description": "p10/p50/p90 in mm/h and P(R > 20/40 mm/h) at the point's Sky pixel"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No run, bundle or city to answer from"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request names a point or cycle that cannot be served"
          }
        },
        "summary": "Rain fan chart at one junction: quantiles and exceedance per step",
        "tags": [
          "nowcast"
        ]
      }
    },
...
    "/v1/pumps/map": {
      "get": {
        "description": "Each assignment of the run's own ``pump_plan.json``, drawn (Jalayantra, SPEC.md 7.6).\n\nPer leg: the depot and the place with coordinates, the plan's ETA, the depth series at the\nplace with and without the pump (recomputed with the optimiser's own functions, with\n``agrees`` saying whether the recount reproduces the plan's minutes), its window above 45 cm,\nand - with ``routes`` - the road a truck would take at the cycle time and what the pump buys\nif it arrives when that road says. Read-only and ungated; the first answer for a run is slow\n(the city's segment table and twelve routes) and is remembered, ``cached`` says which.",
        "operationId": "pumps_map_v1_pumps_map_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "routes",
            "required": false,
            "schema": {
              "default": true,
              "title": "Routes",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pumps Map V1 Pumps Map Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The run's pump plan on the city: depots, roads, and depth with and without each pump",
        "tags": [
          "pumps"
        ]
      }
    },
...
    "/v1/route": {
      "post": {
        "description": "Plan one trip.\n\nBody: ``{origin, destination, depart_at?, profile?, risk_tolerance?, run_id?, spread?,\ntrip_id?, explain?}``.\n\nReturns both routes, because the comparison is the product: a dispatcher who only sees the\nsafe route has no way to judge whether the detour was worth it.\n\n``spread`` (default true) adds ``corridors[]`` - up to three safe roads with the share of\ntraffic the policy gives each, and the one this request is assigned to. ``trip_id`` makes\nthat assignment stable for a trip, so a reader who reloads is not sent somewhere else; it is\nthe client's own random id and nothing is stored against it. ``explain`` (default true) adds\n``reasons[]``, which is structured data - the frontend writes the sentences (TECH_SPEC 3.2).",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route V1 Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route around the forecast water, beside what a naive router does",
        "tags": [
          "route"
        ]
      }
    },
...
  },
...
}